# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Social Media application",
        "title": "Social Media Application API"
    },
    "host": "localhost:3000",
    "securityDefinitions": {
        "jwt": {}
    },
    "paths": {
...
        "/api/likes/{postId}": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like Count of a post.",
                "description": "Get likes count.",
                "operationId": "getLikesCount",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to fetch all comments on it.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In count key of response object you will get likes count of respective postId."
                    },
                    "400": {
                        "description": "Respective post not yet liked. Please verify postId that you want to get."
                    }
                }
            }
        },
        "/api/likes/toggle/{postId}": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Like or Dislike a post",
                "description": "Only authenticated user Toggles Like or Dislike a post with mentioned postId",
                "operationId": "likeOrDislikePost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to fetch all comments on it.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removed like successfully."
                    },
                    "201": {
                        "description": "Liked successfully."
                    },
                    "400": {
                        "description": "Error while like/dislike the post"
                    }
                }
            }
        },
...
        "/api/otp/send": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Send 4 digit OTP to user.",
                "description": "OTP template sending to logged in user on registered email address",
                "operationId": "sendOtp",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Contact details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent to your email id."
                    },
                    "401": {
                        "description": "Error while sending OTP email."
                    }
                }
            }
        },
        "/api/otp/verify": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "OTP verification",
                "description": "Verify OTP generated for registered email address.",
                "operationId": "verifyOtp",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User details to verify otp",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "required": true
                                },
                                "otp": {
                                    "type": "integer",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP verified successfully. Now reset password using reset-password api. "
                    },
                    "400": {
                        "description": "OTP is invalid. Please provide registered email id."
                    },
                    "401": {
                        "description": "User not found! register yourself now!! or Error while OTP verification."
                    }
                }
            }
        },
        "/api/otp/reset-password/{otp}": {
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Reset password with OTP",
                "description": "Reset password using OTP.",
                "operationId": "resetPassword",
                "parameters": [
                    {
                        "name": "otp",
                        "in": "path",
                        "description": "4 digit OTP",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "New password details",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "password": {
                                    "type": "string",
                                    "required": true
                                },
                                "confirmPassword": {
                                    "type": "string",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "You reset password successfully. Login with your new password to continue."
                    },
                    "400": {
                        "description": "Mismatch new password and confirm password! or Error while OTP verification."
                    },
                    "401": {
                        "description": "OTP is expired. Please try again."
                    }
                }
            }
        }
    },
...
}